+++ R/format_table.R
@@ -325,7 +325,7 @@ format_table <- function(x,
   for (stats in c(
     "p_CochransQ", "p_Omnibus", "p_Chi2", "p_Baseline", "p_RMSEA", "p_ROPE",
     "p_MAP", "Wu_Hausman_p", "Sargan_p", "p_Omega2", "p_LR", "p_calibrated",
-    "weak_instruments_p"
+    "weak_instruments_p", "p_Superiority", "p_Inferiority"
   )) {
     if (stats %in% names(x)) {
       x[[stats]] <- format_p(
@@ -673,6 +673,10 @@ format_table <- function(x,
     names(x)[names(x) == "ROPE_low"] <- "ROPE"
     x$ROPE_CI <- NULL
   }
+
+  names(x)[names(x) == "Superiority_Percentage"] <- "Above ROPE"
+  names(x)[names(x) == "Inferiority_Percentage"] <- "Below ROPE"
+
   x
 }